Added a true header adapted from Techmarkets. Will update in another PR to get walletbeacon working
