+++ src/App.tsx
@@ -3,8 +3,8 @@ import './App.css'
 import { ChakraProvider, CSSReset, Stack } from '@chakra-ui/react'
 import { theme } from '@chakra-ui/theme'
 
-import Navigation from './components/Navigation/Navigation'
 import Header from './components/Header/Header'
+import Body from './components/Body'
 import HowTo from './components/HowTo/HowTo'
 import Faq from './components/Faq/Faq'
 import Footer from './components/Footer/Footer'
@@ -13,9 +13,9 @@ const App: React.FC = () => {
   return (
     <ChakraProvider theme={theme}>
       <CSSReset />
-      <Navigation />
+      <Header />
       <Stack minH="90vh">
-        <Header />
+        <Body />
         <HowTo />
         <Faq />
       </Stack>
